+++ CHANGELOG.md
@@ -9,13 +9,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 ### Added
 
-## [Unreleased]
+## [102.1.0] - 2022-10-22
+### Fixed
+- fix infinite loading on submit where returned server error
 ## Added
 - current product default value
 - customer email default value
 - hidden text input type block
 - form identifier parameter for widget
 - create getter methods for success and failure message
+- current cms page default value provider
+- default text defaul value provider
 
 ## [102.0.23]
 ### Fixed